Unrelated: split germany tests into its own module
