+++ workalendar/tests/test_germany.py
@@ -1,0 +1,267 @@
+from datetime import date
+
+from workalendar.tests import GenericCalendarTest
+from workalendar.europe import (Germany, BadenWurttemberg, Bavaria, Berlin,
+                                Brandenburg, Bremen, Hamburg, Hesse,
+                                MecklenburgVorpommern, LowerSaxony,
+                                NorthRhineWestphalia, RhinelandPalatinate,
+                                Saarland, Saxony, SaxonyAnhalt,
+                                SchleswigHolstein, Thuringia)
+
+
+class GermanyTest(GenericCalendarTest):
+    cal_class = Germany
+
+    def test_year_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 1, 1), holidays)
+        self.assertIn(date(2014, 4, 18), holidays)
+        self.assertIn(date(2014, 4, 21), holidays)
+        self.assertIn(date(2014, 5, 1), holidays)
+        self.assertIn(date(2014, 5, 29), holidays)
+        self.assertIn(date(2014, 6, 9), holidays)
+        self.assertIn(date(2014, 10, 3), holidays)
+        self.assertIn(date(2014, 12, 25), holidays)
+        self.assertIn(date(2014, 12, 26), holidays)
+
+    def test_year_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 1, 1), holidays)
+        self.assertIn(date(2015, 4, 3), holidays)
+        self.assertIn(date(2015, 4, 6), holidays)
+        self.assertIn(date(2015, 5, 1), holidays)
+        self.assertIn(date(2015, 5, 14), holidays)
+        self.assertIn(date(2015, 5, 25), holidays)
+        self.assertIn(date(2015, 10, 3), holidays)
+        self.assertIn(date(2015, 12, 25), holidays)
+        self.assertIn(date(2015, 12, 26), holidays)
+
+    def test_year_reformation_day(self):
+        # Reformation Day is included each year in:
+        # Brandenburg, MecklenburgVorpommern, Saxony, SaxonyAnhalt & Thuringia
+        holidays = self.cal.holidays_set(2015)
+        if self.cal.include_reformation_day:
+            self.assertIn(date(2015, 10, 31), holidays)
+        else:
+            self.assertNotIn(date(2015, 10, 31), holidays)
+        # But in the year 2017, it's included for the whole country
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 10, 31), holidays)
+
+
+class BadenWurttembergTest(GermanyTest):
+    cal_class = BadenWurttemberg
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 1, 6), holidays)
+        self.assertIn(date(2014, 6, 19), holidays)
+        self.assertIn(date(2014, 11, 1), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 1, 6), holidays)
+        self.assertIn(date(2015, 6, 4), holidays)
+        self.assertIn(date(2015, 11, 1), holidays)
+
+
+class BavariaTest(GenericCalendarTest):
+    cal_class = Bavaria
+
+    def test_year_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 1, 1), holidays)
+        self.assertIn(date(2014, 1, 6), holidays)
+        self.assertIn(date(2014, 4, 18), holidays)
+        self.assertIn(date(2014, 4, 21), holidays)
+        self.assertIn(date(2014, 5, 1), holidays)
+        self.assertIn(date(2014, 5, 29), holidays)
+        self.assertIn(date(2014, 6, 9), holidays)
+        self.assertIn(date(2014, 6, 19), holidays)
+        self.assertIn(date(2014, 8, 15), holidays)
+        self.assertIn(date(2014, 10, 3), holidays)
+        self.assertIn(date(2014, 11, 1), holidays)
+        self.assertIn(date(2014, 12, 25), holidays)
+        self.assertIn(date(2014, 12, 26), holidays)
+
+    def test_year_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 1, 1), holidays)
+        self.assertIn(date(2015, 1, 6), holidays)
+        self.assertIn(date(2015, 4, 3), holidays)
+        self.assertIn(date(2015, 4, 6), holidays)
+        self.assertIn(date(2015, 5, 1), holidays)
+        self.assertIn(date(2015, 5, 14), holidays)
+        self.assertIn(date(2015, 5, 25), holidays)
+        self.assertIn(date(2015, 6, 4), holidays)
+        self.assertIn(date(2015, 8, 15), holidays)
+        self.assertIn(date(2015, 10, 3), holidays)
+        self.assertIn(date(2015, 11, 1), holidays)
+        self.assertIn(date(2015, 12, 25), holidays)
+        self.assertIn(date(2015, 12, 26), holidays)
+
+
+class BerlinTest(GermanyTest):
+    cal_class = Berlin
+
+
+class BrandenburgTest(GenericCalendarTest):
+    cal_class = Brandenburg
+
+    def test_year_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 1, 1), holidays)
+        self.assertIn(date(2014, 4, 18), holidays)
+        self.assertIn(date(2014, 4, 20), holidays)
+        self.assertIn(date(2014, 4, 21), holidays)
+        self.assertIn(date(2014, 5, 1), holidays)
+        self.assertIn(date(2014, 5, 29), holidays)
+        self.assertIn(date(2014, 6, 9), holidays)
+        self.assertIn(date(2014, 10, 3), holidays)
+        self.assertIn(date(2014, 10, 31), holidays)
+        self.assertIn(date(2014, 12, 25), holidays)
+        self.assertIn(date(2014, 12, 26), holidays)
+
+    def test_year_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 1, 1), holidays)
+        self.assertIn(date(2015, 4, 3), holidays)
+        self.assertIn(date(2015, 4, 5), holidays)
+        self.assertIn(date(2015, 4, 6), holidays)
+        self.assertIn(date(2015, 5, 1), holidays)
+        self.assertIn(date(2015, 5, 14), holidays)
+        self.assertIn(date(2015, 5, 25), holidays)
+        self.assertIn(date(2015, 10, 3), holidays)
+        self.assertIn(date(2015, 10, 31), holidays)
+        self.assertIn(date(2015, 12, 25), holidays)
+        self.assertIn(date(2015, 12, 26), holidays)
+
+
+class BremenTest(GermanyTest):
+    cal_class = Bremen
+
+
+class HamburgTest(GermanyTest):
+    cal_class = Hamburg
+
+
+class HesseTest(GermanyTest):
+    cal_class = Hesse
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 6, 19), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 6, 4), holidays)
+
+
+class MecklenburgVorpommernTest(GermanyTest):
+    cal_class = MecklenburgVorpommern
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 10, 31), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 10, 31), holidays)
+
+
+class LowerSaxonyTest(GermanyTest):
+    cal_class = LowerSaxony
+
+
+class NorthRhineWestphaliaTest(GermanyTest):
+    cal_class = NorthRhineWestphalia
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 6, 19), holidays)
+        self.assertIn(date(2014, 11, 1), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 6, 4), holidays)
+        self.assertIn(date(2015, 11, 1), holidays)
+
+
+class RhinelandPalatinateTest(GermanyTest):
+    cal_class = RhinelandPalatinate
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 6, 19), holidays)
+        self.assertIn(date(2014, 11, 1), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 6, 4), holidays)
+        self.assertIn(date(2015, 11, 1), holidays)
+
+
+class SaarlandTest(GermanyTest):
+    cal_class = Saarland
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 6, 19), holidays)
+        self.assertIn(date(2014, 8, 15), holidays)
+        self.assertIn(date(2014, 11, 1), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 6, 4), holidays)
+        self.assertIn(date(2015, 8, 15), holidays)
+        self.assertIn(date(2015, 11, 1), holidays)
+
+
+class SaxonyTest(GermanyTest):
+    cal_class = Saxony
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 11, 19), holidays)
+        self.assertIn(date(2014, 10, 31), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 11, 18), holidays)
+        self.assertIn(date(2015, 10, 31), holidays)
+
+    def test_extra_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 11, 16), holidays)
+        self.assertNotIn(date(2016, 11, 23), holidays)
+        self.assertIn(date(2016, 10, 31), holidays)
+
+
+class SaxonyAnhaltTest(GermanyTest):
+    cal_class = SaxonyAnhalt
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 10, 31), holidays)
+        self.assertIn(date(2014, 1, 6), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 10, 31), holidays)
+        self.assertIn(date(2015, 1, 6), holidays)
+
+
+class SchleswigHolsteinTest(GermanyTest):
+    cal_class = SchleswigHolstein
+
+
+class ThuringiaTest(GermanyTest):
+    cal_class = Thuringia
+
+    def test_extra_2014(self):
+        holidays = self.cal.holidays_set(2014)
+        self.assertIn(date(2014, 10, 31), holidays)
+
+    def test_extra_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 10, 31), holidays)
